requirements.txt
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
EbookLib==0.20
idna==3.11
lxml==6.0.2
pathvalidate==3.3.1
python-magic==0.4.27
requests==2.32.5
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
